feat: migrate first 2 apply prompts to dedicated module - Created vibectl/prompts/apply.py with PLAN_APPLY_PROMPT and apply_output_prompt - Removed migrated prompts from main prompt.py - Updated imports in apply_cmd.py, execution/apply.py, and test files - Added comprehensive tests with 100% coverage - All tests passing
