+++ PLANNED_CHANGES.md
@@ -25,11 +25,15 @@ Refactor the `vibectl apply` implementation to bring it in line with the `vibect
 
 ## Planned Refactoring Steps
 
-### Phase 1: Extract Execution Logic
-- [ ] Create `vibectl/execution/apply.py` module
-- [ ] Move `_run_intelligent_apply_workflow()` and related functions from `apply_cmd.py`
-- [ ] Update `apply_cmd.py` to use the extracted execution module
-- [ ] Ensure all existing functionality is preserved
+### Phase 1: Extract Execution Logic ✅ COMPLETED
+- [x] Create `vibectl/execution/apply.py` module
+- [x] Move `_run_intelligent_apply_workflow()` and related functions from `apply_cmd.py`
+- [x] Update `apply_cmd.py` to use the extracted execution module
+- [x] Ensure all existing functionality is preserved
+- [x] **CLEANUP**: Remove dead code and unused imports from `apply_cmd.py`
+- [x] **CLEANUP**: Achieve 100% test coverage for simplified `apply_cmd.py`
+
+**Phase 1 Summary**: Successfully extracted 875 lines of execution logic to `vibectl/execution/apply.py`, cleaned up 985 lines of dead code from `apply_cmd.py`, and achieved clean separation of concerns. All tests passing with 100% coverage for the command handler.
 
 ### Phase 2: Extract Prompts
 - [ ] Create `vibectl/prompts/apply.py` module
